@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import '../css/Welcome.css';
 import {AuthenticationCredentials} from './elements/authentication';
-import {name, version} from '../../package.json';
+import packageInfo from '../../package.json';
 
 /**
  * Welcome - the Getting started component.
@@ -100,7 +100,7 @@ const Welcome = (props) => {
         <a className='open-source' onClick={openIssues}>
           Contact us
         </a> with your feedback.<br/>
-        {name} v{version}
+        {packageInfo.name} v{packageInfo.version}
          - Copyright © 2021 <a className='mcin' onClick={openMCIN}>
         MCIN</a>.
       </div>
